+++ mmeval/__init__.py
@@ -2,7 +2,6 @@
 
 # flake8: noqa
 
-from .classification import *
 from .core import *
-from .segmentation import *
+from .metrics import *
 from .version import __version__
