+++ mmeval/metrics/__init__.py
@@ -1,5 +1,6 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 
 from .accuracy import Accuracy
+from .mean_iou import MeanIoU
 
-__all__ = ['Accuracy']
+__all__ = ['Accuracy', 'MeanIoU']
