--- mmeval/segmentation/__init__.py
@@ -1,5 +1,0 @@
-# Copyright (c) OpenMMLab. All rights reserved.
-
-from .mean_iou import MeanIoU
-
-__all__ = ['MeanIoU']
